BAVL-115: Moved entity/model mapping into service layer modules
